🔨 [refactor] 디렉토리, 불필요한 오버라이드 리팩토링
